hyperv: Introduce HV_ENABLE_VP_VTL hcall
Introduce the Hyper-V VSM HV_ENABLE_VP_VTL hypercall. This hypercall
configures an initializes the CPU state of a VTL. A new VpVsm object is
created to keep track for the VTL configuration, as well as helpers to
extract the VTL level from a CPU's APIC ID.

The initialization goes through the following steps:
 - Initialize a new QEMU CPU with the same physical APIC ID as the
   caller, and an APIC group that matches the VTL.
 - Attach VpVsm object to that vCPU.
 - Set initial CPU state based on the hypercall input.

The VTL CPU doesn't yet run, this happens on the first VTL call
hypercall.

Signed-off-by: Nicolas Saenz Julienne <nsaenz@amazon.com>