kvm: Introduce poll() support to vCPU thread
Introduce the option to poll a vCPU when it's stopped after a KVM exit.
While polling, if KVM event becomes available to that vCPU, it'll return
into QEMU, and call the vCPU's poll_callback(). The callback can be used
to restart the vCPU's execution or to inform other threads that the vCPU
has events pending.

This feature can be built upon to implement halting in user-space as
well as to model Hyper-V Virtual Secure Mode VTLs.

Signed-off-by: Nicolas Saenz Julienne <nsaenz@amazon.com>